ios [nfc]: Move NotificationTapEventListener to its own Swift file
